backend/requirements.txt

Flask==2.3.3
Flask-Cors==4.0.0
numpy==1.24.3
opencv-python==4.7.0.72
tensorflow==2.12.0
keras==2.12.0
mysql-connector-python==8.0.33
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.0
